Enable playbin support in Goo, rclark (ACT_A0873619-PLAYBIN-20090225)
changes to support pipelines constructed by playbin (even in tunneled mode).  For example from the command line:
   gst-launch playbin uri=file:///tmp/StarWars_V-H264_1.5Mbit_VGA_30fps_A-AAC_48kHz@128kbps.3gp
Playbin will figure out the proper demux/decoders/etc to use based on the file (or stream) type, and the type of the contained audio and/or video streams. This also enables the GUI video player in poky to function, as well as pretty much every other gstreamer based A/V player in existence.