fixes in dasfsink
Even though we bypass the ringbuffer functionality of the base class, we still need to pretend and maintain the state which is normally maintained by the ringbuffer.  Without this, the thread that pushes buffers to render will keep pushing audio buffers to render (while holding the preroll lock, which blocks the thread which is trying to propagate the PLAYING->PAUSED state change to the rest of the elements in the bin).

There is still an issue with state changes, as the dasfsink gets left in an asynchrounous state change..  so a PLAYING->PAUSED followed by a PAUSED->PLAYING fails (but if you put a seek in the middle, that somehow resets the dasfsink's state... I'm still investigating that)