gstgoovideofilter/gstgooaudiofilter: defer sending OMX_BUFFERFLAG_EOS to OMX until the sink receives the GST_EVENT_EOS
Buffers can be queued up in GstBaseSink, meaning that the video/audio decoder/encoder will see the GST_EVENT_EOS while there is still data that hasn't reached the sink.  The problem is worse with playbin, as there can be a GstQueue between the decoder and sink.  The way this is handled is that when the GstGooVideoFilter/GstGooAudioFilter's outport is tunneled, it waits to send the EOS to OMX until it gets a "reverse-eos" event from the sink.  The GstDasfSink and GstGooSinkPP are modified to send this custum event back upstream when they get the GST_EVENT_EOS.