gstgooutils: fix some memory leaks
The most important part is that I needed to call gst_iterator_free() in a couple spots.  But there also were a few other cleanups, in particular simplifying the logic when we meet the return conditions in the middle of iterating a list (so we don't forget to unref 'item').  And another change to reduce the number of times we call the pipeline-change callback (which does fix any problems, but is a bit more efficient and more importantly reduced the number of unrelated traces to look at while hunting down the ref leak).