gstgoovideofilter: fix for high CPU load with camera pipeline
the basic issue:  In GstBaseSrc, there is a thread that basically does
roughly the following (psuedocode):

    while(TRUE) {
	self->create(&buffer);  // calls gst_goo_camera_create()
        gst_pad_push(buffer);   // calls gst_goo_video_filter_chain()
                                // which returns immediately
    }

So this thread is basically pushing ghost buffers as fast as possible,
with no connection to the actual OMX buffers that are flowing down the
pipe at the OMX level.  If I do a camera-only pipeline, this is
happening every 2 to 2.5ms, and CPU is at 100%.

With this patch, there is a semaphore in GstGooVideoFilter, which is
released (incremented) in _outport_buffer() function, and obtained
(decremented) in the _chain() function.  This will cause the
gst_pad_push() in the above described loop to suspend until the buffer
comes through the OMX pipe.  With this patch, I can get 30fps in a
camera-only pipeline with ~20% CPU utilization.   (There are probably
other optimizations possible too.. but that can come later.)

The problem with this approach is that it only works if you don't put
a GstQueue between the GstGooCamera and the encoder.  And I'm not
really sure what would happen in a scenario with no encoder.  So I'm
trying to figure out if there is some way to do this in GstGooCamera
independently of the encoder.