gstgoodecjpeg: optimizations for header processing
In non-packetized mode, when we don't get an entire jpeg img in a single GstBuffer, defer the header processing until we get EOS.  Otherwise we could attempt parsing all the headers for each 4kb of data (which could translate into many 100's of times per image) which becomes quite expensive.

Note:  if using multifilesrc, be sure to head the comment at the top of the gstmultifilesrc.c and be sure to specify a 'framerate' parameter in the 'caps' property.  This will ensure that gstgoodecjpeg (or any img decoder element) goes into packetized mode.