fix failed assert (G_IS_OBJECT()) and hang at end of playback
Issue is that goo component gets finalized while someone else still holds a reference.  There are a number of other places where the ref_count is manipulated directly, which should be removed.. but I will make that a different commit so we can merge that when we have sufficient time to test.