ARM_AAC Decoder Fix
This patch allows to decode correctly eAAC+ files comming from a demuxer. Some demuxed files does not report number of channels correctly, it could be set as 1, this will result in a distorted playback. It always be set to 2 channels. Also included is a fix for AAC-HE coming out from a demuxer. For audio only eAAC+ files, the sample rate was adjusted.