A better way to send messages to the bus from the camera component.
When camera moves from one state to another, the camera now sends
messages indicating the state transition and the timestamp. Specific
purposes applications, like 'shot' can catch these in order to
measure: shot to shot (pause->playing->pause) and stanby to first shot
(null->ready->pause->playing). shot to snapshot can be computed
but the end-point is inside the omx layer, so the PP should implemented
this point. shot to save can also be computed, but the end-point is
inside the multifilesink component/file so the latter should be also
implemeted.