Add DASF support in Audio Encoders.
Several changes done to the dasfsrc element in order to support recent implementations of audio encoders, namely PCM, AAC and NB-AMR Encoders.

Improved code readability in dasfsrc plug-in.

GstGooAudioFilter is now the parent of PCM encoder. This is done in order to use the current dasfsrc implementation, which assumes an audio filter class as its next element

Correct timestamp and duration for GstBuffers in DASF mode for AAC and NB-AMR encoders.

Signed-off-by: Daniel Diaz <ddiaz@ti.com>